Recognize (and skip) keyword 'RPTSCHED'
Otherwise, it is possible for the reader to become confused if an
input deck contains a record of the form

    RPTSCHED
    WELSPECS WELL /

which would trigger the read for 'WELSPECS'.

This is a proper reimplementation of commit 1f4e881 which used
functionality that is not available in master.

Point hat: Bard.Skaflestad@sintef.no