plotToolbar: Add support for 1D plotting of datasets if the dataset.
This commit introduced additional button controls to plotToolbar for when the data can easily
be visualized as a series of line plots instead of a full blown patch. Mostly useful for 1D grids,
so the option is only visible if the grid has only one direction with more than one cell (according
to G.cartDims) or cartDims is missing.