plotCellData: Handle Special Case of Single Element of Scalar Field
Function plotCellData failed to handle the case of selecting a single
element from a scalar (indexed/interpolated) grid vector.  This case
would most commonly occur when restricting the view to a single cells in
a two-dimensional model (either strict 2D or topSurfaceGrid), but can
happen in other cases too.

This commit uses the current figure's colour map and the input data's
value range to translate the single data value to an RGB triplet that is
passed on to function plotPatches.