Adding a "validateState" routine to AD framework. This routine is called on state at the beginning of all simulations.
The routine will give errors for missing or malformed fields (e.g. wrong number of saturation columns) and can fix
minor issues (e.g. ".rs" missing for a black oil model when disgas is not enabled) that would otherwise spoil a
perfectly good simulation. This is in preparation of the MRST 2016a release.

Currently implemented for blackoil (and derivatives), surfactant, polymer and compositional