Core: Add utility for downloading fileset from GitHub
This commit adds a function, githubDownload, that will download,
without history, either a specific list of files or all files within
a public GitHub repository.  This is mainly useful when accessing
public datasets that are hosted on GitHub such as the "OPM-Data"
repository.  The caller can request a specific revision (i.e.,
commit or tag) or default to the symbolic revision 'master'.  By
default, output is placed in a subdirectory of 'mrstDataDirectory'
although the user can override this setting.

Note: using WEBSAVE is a fairly heawy-weight approach and we should
probably investigate using the GitHub services API instead, but for
now this is enough to access some interesting datasets.

The following statements--from the function documentation--downloads
the files

    * SIMPLE_SUMMARY.INIT
    * SIMPLE_SUMMARY.UNRST
    * SIMPLE_SUMMARY.UNSMRY

from the "flow_diagnostic_test/eclipse_simulation" collection of the
OPM/opm-data repository tagged at revision 2198d5b.

    repo     = 'OPM/opm-data';
    rev      = '2198d5b';
    filebase = 'flow_diagnostic_test/eclipse-simulation';
    ext      = { '.INIT', '.UNRST', '.UNSMRY' };
    files    = strcat('SIMPLE_SUMMARY', ext);

    files = githubDownload(repo, 'Revision', rev, ...
                           'File', strcat(filebase, '/', files));