Dataset Path: Add option to bypass availability check
The new option, 'skipAvailableCheck', enables manual operation on a
known dataset directory without checking if the dataset is available
or downloading if not.  The typical example us the case of a dataset
for which the 'fileurl' is unknown--e.g., the Norne simulation model.

Note: Since this option bypasses the safety measures of function
'getDatasetPath', the option should only be used in constrained
situations.