readGRDECL: Customisable treatment of missing INCLUDEs
This commit adds a user hook for treating missing INCLUDE files,
i.e., resources referenced in an INCLUDE statement that are
nevertheless missing from the on-disk representation of the case.
By default, function 'readGRDECL' will fail (i.e., call 'error') if
we encounter a missing file but the user may pass a function like

    @(varargin) []

in which case missing files will be ignored (meaning completely
skipped).