DownloadMBGL: Reflect New Location in 'mrstPath'
Previously we would only write the new location to persistent
storage (fullfile(ROOTDIR, 'startup_user.m')), but not update the
active module mapping.  While the modload script would nevertheless
do the right thing, it is better to close the semantic hole--not
least because it makes the path display reflect reality.

While here, and in the interest of MRST 2016b release preparation,
also verify that the file exchange URL still points to a valid
location.