Add Support Two-Phase Oil/Gas Simulations
This commit introduces a new class,

    TwoPhaseOilGasModel < ThreePhaseBlackOilModel

that implements its equations in terms of 'equationsBlackOil'.  We use
Model 1 from the Tenth SPE Comparative Solution Project as a simple
example of this support.  The example is simple because the fluid
properties are mostly trivial (immiscible, incompressible, essentially
quadratic relative permeability curves).

The changes to 'equationsBlackOil' are substantial.  We may want to
rework this in terms of a dedicated 'equations*' function for the 2P OG
system.  For now, this is submitted for wider testing.