mrstModule: Adding behaviour mode.
This adds an intentionally undocumented option for setting the
module manager into either release or experimental mode. The default
is experimental mode which behaves identically to earlier versions
of mrstModule. When experimental mode is set, all folders named
experimental are excluded. This makes it easier to ensure that new
modules do not depend on internal code we are not actually releasing.