Incomp Well DP: Account For Possibility Of Mismatching Array Sizes
We have several examples that have mismatching number of fluid phases vs
number of well injection components (i.e., W.compi).  While we should
really fix that Soon[tm], there is not enough time before the release of
MRST 2016b to do the change correctly.  As an interim measure, switch to
using the minimum of the number of fluid phases and numel(W.compi) as
the number of array components to consider when computing 'rhoMix'.

Also, to guard against this simply becoming another convenient behaviour
we issue a (rate limited) warning in case the number of fluid phases do
not match in any particular call to 'computeIncompWellPressureDrop'.