Update getAtlasGrid to avoid reading any readme.txt files
This update is required to avoid an error that occurs when calling
grdecl = getAtlasGrid() or grdecl = getAtlasGrid(pn1, pv1, ...), now
that README.txt files exist in some of the Atlas data folders. This
error is not encountered by any current co2lab example, however this
update preserves the option to call gr = getAtlasGrid(), etc.