Add Sleipner model and plumes to dataset manager
The actual Sleipner model is not the epi server, rather the
dataset manager refers to the webpage where is can be applied for.

The Sleipner plumes are now on the epi server. These are .mat files that
were created by reading from color illustrations in published papers.
Reference is made to these papers.