Remove sealing formations from list
Not and Ror formations are considered to be sealing formations, and are
not considered for CO2 storage according to the Norwegian Petroleum
Directorate's CO2 Storage Atlas. We remove it from the list of potential
formations to explore.

The Nordmela formation is also sealing, however we
leave it in the list of possible formations to explore because the
permeability and porosity of this formation is explicitly given, so it
may be interesting to explore.