recovering blackOilPolymerExample running
by commenting out the wellVarNames in ThreePhaseBlackOilPolymerModel.

With this way, the polymer model will still use the old well model. Will
investigate whether it is not easy to incorporate the new well model
with all the polymer related features. More specifically, the shear
effects and the mixing parameter for the wells.