ADI: Avoid slightly optimistic use of reduceToDouble.
This was causing issues in expressions of the form
var = double2ADI(var, sample)
var(1) = somead;
var(2) = someotherad;

Since the first call might not actually make the variable AD.
This is a misbehavior inherit to Matlab's way of casting that we
should not mess with. Instead, introduce reduceToDouble as member
function and add an equivialent dummy function for doubles so that
this process can be done manually in places where it is known to
be safe.