makeLayeredGrid: Add Support for Variable Layer Thickness
This commit extends function 'makeLayeredGrid' to support layers
of variable thickness.  If the user passes a vector of real,
hopefully positive, numbers as the second parameter, then we now
interpret that as a vector of layer thicknesses.  We still treat
a single scalar as a number of layers to create in order to
preserve backwards compatibility with existing callers.  This
means that it is not possible to create a non-unit thickness
single layer without additional manual steps, however.

Add a few examples to illustrate the extended capabilities.

Thanks to Jhabriel Varela on the MRST user forum for pointing
out the obvious extension to usability.