Modify optimizeRates.m to handle separate scaling of well controls
These changes to optimizeRates.m are required when a well control is
'bhp' during the injection period, and then 'rate' during the migration.
There is most likely a more elagant way to implement this separate
scaling of well controls for two different control periods, however this
commit is a first attempt to do so.