Deck Reader: Add Support for Dual Porosity
In an ECLIPSE DP model the matrix is the first half of the cells
in th k direction and the fracture cells are the second half of
the cells in the k direction. Essentially the number of cells in
the vertical direction is double for ECLIPSE DP.

Essentially reads in one grid but assigns two properties to each
cell instead of one.

Properties (PERMX, PORO, SATNUM etc) are read in two coloumn
format--(noCells, 2).  The first column is the matrix properties
and the second column is the fracture properties.

Two rock structures and two fluid structures are created.  One
for the matrix (using column 1 data from properties in the deck)
and one for the fracture (using column 2 data from properties in
the deck).

Unit Conversion have been added for the new DP_Info output, this
includes the shape factor, block dimensions, SIGMAV.

The process wells is crudely implemented.  If there is a
perforation in the fractures (z>cartDims(3)) then the z is
changed to

   znew = z - cartDims(3)

and that perforation is labelled as fracture otherwise the
perforation is in the matrix and the z stays the same but that
perforation is labelled as matrix.  Two wells are created--one
in the fractures and one in the matrix.  The processWellsDP is
not very robuts so I call it outside of the initEclipseModelDP
function.

I have not extended the equilbrate state to dual porosity.