AMGCL: Prepare for MSVC (Windows) Build Support
This commit prepares the automated build script 'amgcl_matlab' for
adding the requisite incantations to support the MSVC compiler on MS
Windows.  Most importantly, we switch to using the new MEX syntax
for passing flags and as a result the build script no longer works
on MATLABs prior to 8.3.0 (R2014a).