IncompTPFA: Support Capillary Pressure in Presence of Fractures
Function 'capPressureRHS' was not aware of the possibility of
non-neighbouring connections arising from fracturing processes.
This commit switches to using the 'getNeighbourship' helper
function with 'Topological' set to TRUE whence additional
connections across fractures or internal boundaries are included
in the neighbourship definition.

While here, also cater to the possibility that we pass actual
transmissibility values to the internal helper function getFlux
of twophaseJacobian.  The previous implementation assumed that
we would only pass one-sided (half-) transmissibilities to this
helper function.

Collectively, these changes enable running the

    'basic_model_CPC'

example of Gallileu Genesis--posted to the MRST User Forum

https://groups.google.com/forum/#!topic/sintef-mrst/7wy4w1JL8zw

I have verified that all existing examples that use 'incompTPFA'
and/or twophaseJacobian continue to run as expected.