readVector: Use TEXTSCAN to Ameliorate Repeat Count Cost
This commit reimplements the deckformat module's readVector
function in terms of TEXTSCAN as opposed to a home-made scanner
held together with string and chewing gum that was only needed
to support versions of Octave that did not have TEXTSCAN.

The primary motivation for this reimplementation is a user
report from Victoria Spooner (Heriot-Watt University) that
demonstrated unacceptable read performance on a large model
(more than 5 million cells) that featured a ZCORN vector with
repeat counts.

This readVector implementation uses two stages, a simplified
first stage that tries to read a vector for which each element
is listed separately.  This is the common case, especially for
the large vectors such as ZCORN, so it makes sense to optimize
for this case.  If this scanner fails--typically because we
encounter a repeat count character--then we switch to the more
elaborate read scheme.  At this stage we trade memory for speed
and read all tokens (separated by white-space) into a cellstring
and then perform a second pass over this array to decode the
repeat counts and convert the element strings to floating-point
numbers.

Finally, we execute a little cleanup routine to consume the
vector terminator character ('/') and to determine whether or
not we successfully input the entire vector.