mrstPath: Guarantee Fully Resolved System Paths for Modules
This commit ensures that system paths added to the module
registry are always fully resolved (i.e., do not contain dotdot
components '..').  This, in turn, means that when a module is
activated through 'mrstModule add' it will use fully resolved
system paths.

Based on pre-release testing by [at]moyner and myself, MATLAB
R2018a's 'path' function appears to always return fully resolved
system paths in which case 'mrstModule's 'prune_modules' helper
function will often think that a module is not activated despite
its directories being present in 'path'.

Note that this is sort of a hack, but it is the least invasive
change we can make in anticipation of MATLAB R2018a which will
be released sometime between MRST 2017b and MRST 2018a and which
user's may start using before the subsequent MRST release.