Add Debugging Utility for MRST Path Issues
This commit introduces a new utility function, mrstExtraDirs, that
will extract a list of directories on the MATLAB search path that do
not come from MATLAB itself but are typically added by MRST's core
or some of its add-on modules.  I have found this to be invaluable
when debugging path-related problems.