addWell: Add Option to Disable 'rR' Calculation
This commit introduces a new option, calcReprRad, defaulting to
TRUE, that specifies whether or not to calculate representative
radii for each perforation/completion.  If set to false by
explicit action at the call site, then the representative radius
field will be empty and the resulting well structure is not
usable for polymer injection.  This is nevertheless useful when
*NOT* using the polymer injection capability--especially in the
context of a simulation graph grid that does not meet all of
geometric requirements of the 'radiusRep' calculation.