ReservoirModel and derived classes: Finally, get rid of saturationVarNames and componentVarNames.
These fields have been deprecated in favor of getters for a while now, which is easier for dynamic models.
This change may break a few constructors here and there. I do not think we have any code that relies on
having alternate phase ordering (which is no longer supported but was included in the first AD-OO version
where everything was completely flexible)