ad-core: Fix missing processing of deck for operators
This fixes a bug where inherited models may not have access to the deck directly.

This bug was present due to fixes to remain compatible with Octave. In this commit,
I try to detangle the deck-related features from the constructor. Not entirely satisfactory, in that we still need to manually call setupOperators to get the right
behavior if we are not using selectModelFromdeck.

Pointy hat: I think both me, Matlab and Octave is to blame for this one.