Output/Data Directories: Create Defaults Only if Needed
This commit ensures that we do not create the default output or data
directories

    fullfile(ROOTDIR, 'output')
    fullfile(ROOTDIR, 'examples', 'data')

if the user of functions mrstOutputDirectory and mrstDataDirectory
pass directory strings as arguments to the initial calls to these
fucntions (i.e., when the persistent variable is empty).  We
nevertheless ensure that the initial directory exists before we
return from the initial call to the *Directory functions.

While here, also exclude fullfile(ROOTDIR, 'output') from Git's file
discovery mechanism.