GRID Section: Parse PINCHREG Keyword
This commit adds a simple parser for the PINCHREG keyword (variable
PINCH specification per region implied by PINCHNUM or FLUXNUM).  The
most finicky parts of this is the expected number PINCREG records.

If the deck uses PINCHNUM (the usual case), then the number of
records is NRPINC (item 3 of GRIDOPTS).  Otherwise, if the deck uses
FLUXNUM, then the number of records is NTFREG (item 4 of REGDIMS)
or, if specified the maximum of NTFREG and MXNFLX (item 11 of
TABDIMS).

This commit mostly punts on these issues, but will take appropriate
action if the *NUM vectors are specified before encountering keyword
PINCHREG.