processWells: Change composition for "generic" producers
These producers (bhp, thp) can be used for any combinations of fluid models. We cannot guarantee that the oil-phase will be present for these models. For this reason, we change the values to be 1 everywhere so that the sum will always be non-zero as long as one phase is present.