ADI Fluid: Recognize GRAVITY Keyword
This is an alternative specification of fluid component mass density
at surface conditions.  The data elements are pure scalars that
define the specific gravities of oil, gas, and water relative to
mass densities of pure water and air at standard conditions which in
E100 is 1 atmosphere and 60 degrees Fahrenheit (15.56 Celsius).