ad-core: Boundary conditions support for multiple phase pressures
This commit adds the possibility of bc.value having n rows, where n
is the number of phases. This allows for a simple implementation of
capillary forces. It also allows for specifying cross-flow for each phase,
as well as a number of strange combinations of sat, mob and value.

Some care should be taken if this feature is to be exposed in addBC.