processWells: Reduce Cost of Cartesian -> Active Cell Mapping
This commit uses the fact that the grid, and therefore its set of
active cells, is constant in a call to 'processWells'.  Therefore we
can precompute the reverse mapping (Cartesian/global cells to active
cells) and pass that as a parameter to the low-level 'buildWell'
and 'active_perf' helper functions.  That, in turn, bypasses a call
to global function 'cart2active' for each record in the WCON*
keywords.  Those can add up to considerable costs--especially for
large models.

Suggested by: [at]steink