Well Keywords: Handle Multiply Specified Wells
This commit adds a measure of robustness to the parsing of WCON*.
The new handling detects the case of multiple records within a
single keyword applying to the same well (or well list).  In this
case we issue a diagnostic of the form

  Warning: WCONINJE: Multiple Records Apply to Same Well/Well-List.
    -> Using Last Entry for { 'C-4H' }.

and, as indicated, use the last entry/record for the affected wells.