AD: Split convergence check into a subroutine for getting tolerances
We are backwards compatible in terms of interfaces with this change. However, dealing with "closeness to convergence" is significantly easier since the underlying tolerances are now accessible together with the values they are checked against.

This is needed to treat linesearch rigorously and also to allow for "acceptable" minimum convergence.