NonLinearSolver: Adding acceptanceFactor property
acceptanceFactor is an additional parameter which makes life easier when "almost good enough" is better than cutting the timestep.

Specifically, if the nonlinearsolver is stuck, then checking if the residuals are less than the tolerance multiplied by some factor is used instead of cutting the timestep. This is disabled by default.