Well Index: Guard Against Non-Finite Ratios
This commit adds a set of robustness fixes to calculating the
connection transmissibility factor (well index) from ratios of
floating-point values (permeability values and the numerator and
denominator of the effective radius).

In particular, this ensures that the well index will be zero in
a cell of all zero permeability.  That in turn enables running
models that would previously require special case handling to
avoid non-finite entries in the fundamental systems of
simultaneous linear equations.