Read Record String: Add Support for Record Containing '/'
The PATHS keyword legitimately contains '/' characters (within
quoted substrings), so we cannot assume that the first occurence
of the '/' character is the record terminator.  Revise the
implementation of the record reader to only consider the *LAST*
UNquoted substring of a line when determining whether or not a
record is complete.

Also, to handle slashes occurring *after* the record terminator,
ensure that we discard all character data following the first
slash in the last unquoted substring.