Deck Reader: Implement Support for Path Aliases (PATHS keyword)
This commit introduces support for pathname aliases defined by
the PATHS keyword, e.g.,

   PATHS
      ECLPATH '../include/' /
   /

which can subsequently be used to shorten pathnames:

   INCLUDE
      '$ECLPATH/grid/grid.inc' /

We use a simple loop over regular expression tokens to affect
the actual path alias substitution in 'readEclipseIncludeFile'.

Accessing the pathname aliases means that the INCLUDE handler
must become aware of the contents of the RUNSPEC section, so add
an extra input parameter to 'readEclipseIncludeFile' and update
all callers accordingly.