Grid Arrays: Control Initialisation Values
This commit, prompted by a real field case, refines the input of
arrays defined on global grid cells.  We extend the helper function
'readGridBoxArray' to accept a single initial value that is expanded
to full array size if the array does not already exist.  This is
needed to avoid undesired--and undetectable--behaviour that would
arise from initialising multiplier arrays (e.g., MULTZ) to zero
instead of the more reasonable value one.