readVector: Handle Mixture of CRLF with LF in Text Files
This commit makes the call to input routine TEXTSCAN more robust so
that it handles an arbitrary mix of CRLF and LF as line separators.
In particular, the 'MultipleDelimsAsOne' option treats CRLF as a
single token in a file that is otherwise LF-based and the WhiteSpace
option ensures that we don't try to interpret CR as floating-point
value during simplified vector input.

Collectively, these robustness fixes restores our ability to input
the W model without breaking the full-field Norne dataset that
includes VFP tables (https://github.com/OPM/opm-tests/tree/master/norne).