RUNSPEC: Ensure Existence of Certain *DIMS Specifications
There are cases, mostly arising in academia, for which the dimension
specification--especially TABDIMS--is omitted.  In this case we are
supposed to use the default dimensions which specify a single
saturation region, a single PVT region and a single FIP region.

Ensure that we support this situation by always creating a TABDIMS
array containing the default dimensions.  We nevertheless always
defer to the case itself if it provides explicit values for these
items.

Do the same for WELLDIMS, mostly for completeness.  MRST's reader
does not currently use any information in WELLDIMS, but having these
values is useful for consistency checking.