MRST Module: Detect Duplicated Module Names in 'add'
This commit makes function 'mrstModule' inspect the module list
being passed as an argument to 'add' and 'reset', to prune any
duplicated entries.  If there are duplicates, then 'mrstModule' will
respond with a diagnostic of the form

    mrstModule add mrst-rexxport ad-core ad-props ad-blackoil ...
                   deckformat mrst-rexxport ad-props
    Warning: Encountered Duplicated Modules in 'mrstModule' Command Verb 'add' (Using Last Duplicated Entry):
      * ad-props
      * mrst-rexxport

and pick the last of each duplicated entry (as if used in separate
'mrstModule add' calls).

Suggested By: Xavier Raynaud