Added a patch to prevent failing call to lineSearch for small gradients
The 'lineSearch' function removes components of the direction vector
smaller than sqrt(eps) when computing max steplength.  This happens in
the helper function 'alphaMax'.  If all components of the direction
vector is smaller than sqrt(eps), 'lineSearch' fails.  To prevent a call
to 'lineSearch' for such small gradients, I added a patch to the
'getSearchDirection' helper function of 'unitBoxBFGS'.  This patch makes
the algorithm behave similar for very small direction vectors as it already does
in the case where the direction vector has a negative scalar product
with the function gradient.