Plot Faces: Refactor Dimension/Coarse Grid Logic
This commit splits the dimension cases into separate helper
functions in order to make the main logic easier to follow.  We
furthermore split the coarse grid handling of 3D grids out to
another function in the same interest.

While here, also disable a couple of helper functions that aren't
actually used in the data plotting.