pollock: Move a few utilities out from main routine
We move out the interior flux and neighborship definitions. These are static for G+state and G respectively and can be very expensive to recompute for if multiple calls for pollock are required for a single grid. The output from these functions can be passed in through varargin, although it is intentionally not documented at the moment.