EOS/flash: A few careful updates to avoid overwriting K-values when transitioning to single-phase state
While here: Document and clean up a bit. We are more careful with the K-values in MRST now than in the earliest implementations since they represent significant effort on part of the EOS. This means that we must take care not to overwrite them with trivial values in the flash.