Capture Changes to Report Settings
This commit makes the ECLIPSE input system aware of the report
settings keywords RPTRST, RPTSCHED, and OUTSOL.  We do not
currently do anything with this information, but having it
available may be useful.

We add parsing support for the older keyword 'OUTSOL' as a means
to being able to parse the NPD_Bjarmeland dataset from the
Norwegian Petroleum Directorate.