Dataset GUI: Support Additional Notes if Applicable
This commit extends the dataset GUI to display additional notes that
may apply to individual datasets--e.g., instructions to slightly
modify input files in order to read them in 'readEclipseDeck'.

Normally there should not be any additional notes so the default
'note' field in the return value from 'datasetInfoStruct' is empty.