Update Reference if INCLUDE Defines at Least One Complete Control
We failed to identify the situation in which an INCLUDE file would
define at least one complete element of 'deck.SCHEDULE.control'.
For a long time we didn't see this problem because all simulation
cases we got used a setup like

    RUNSPEC
    ...
    SCHEDULE
    INCLUDE
      'SCHEDULE.INC' /
    END

On the other hand, the existing implementation breaks in cases with
a setup akin to

    RUNSPEC
    ...
    SCHEDULE
    INCLUDE
      'HISTORY.INC' /
    INCLUDE
      'PREDICTION.INC' /
    END

and especially if all 'WELSPECS' data is in 'HISTORY.INC'.

This commit saves the control ID ('cno') before processing an
INCLUDE directive and updates the base control structure ('ctrl')
upon return from INCLUDE if the new control number exceeds the
original or if the INCLUDE ended while still defining a set of
controls.