Try to Open INCLUDE File Case Insensitively on Unix If Initial Open Fails
This commit adds an expensive fall-back routine for MRST's INCLUDE
file handling on Unix/Linux (i.e., on case sensitive filesystems) if
the initial FOPEN call fails.  In this case we will try to open the
file using case-insensitive string matching to allow models created
on case-insensitive (Windows) or case-preserving (macOS) platforms.

We would prefer not to have this as it is essentially a measure of
desperation, but adding this option allows us to read NPD_Frigg on
Linux systems.