GitHub Download: Fix Filemove for 'All' Files
This commit adds a special-purpose file mover from 'base' to 'dest'
that applies only in the case of 'all' files.  In this case, there
will be a subdirectory named

    [repo, '-', revision]

in FULLFILE(base, user, repo).  The new file mover identifies this
subdirectory and explicitly moves the contents of this directory
into the destination directory (opt.dest).  While here, also reduce
the scope of various diagnostic related variables for future
maintainability.