refactoring WellPhaseFlux.m a little bit
hopefully, it makes the further refactoring easier.

There is one accidental finding is that, by putting content into
function calculatePhaseRate, it avoids to modify the mobw twice from the
total mobility mobt. Most likely, it only matters for the first
iteration, while it is something we need to be careful about.