Promote MEX Flag Genration Function to MRST-Core
This commit adds the function

    mrstDefaultMexFlags

into MRST-Core's set of utility functions.  This function, which
originated as

    setupMexOperatorBuildFlags

in the 'ad-core' module, is a simple way of creating a common
set of compiler/linker/library flags that many (most) of MRST's
MEX functions need.  In particular, this is a way of accessing
MATLAB's built-in BLAS/LAPACK libraries as well as MATLAB's
bundled copy of the Intel OpenMP runtime support library if
compatible with the toolchain used to build the MEX function.