Save and Restore 'mrstPath' Settings for Background Simulation
This commit adds a preprocessing step that captures the current path
mappings in the session's 'mrstPath' and forwards those to the
background simulation process through a separate MAT file.  This MAT
file is then loaded as the first step in the background process run
script and entered into that process' mapping table using

    mrstPath 'reregister'

This ability is useful if 'startup' does not establish all of the
requisite module path mappings.  That, in turn, may be the case if
some mappings are not listed in 'startup_user' or if the user is
temporarily experimenting with an alternate version of a module in a
non-default location.