Process GRDECL: Add Support for Capturing Original CP Vertices
This commit implements a new option in the processGRDECL grid
construction function.  Option 'PreserveCpNodes', a boolean
value whose initial default value is FALSE, specifies whether or
not to identify the original corner-point vertices derived from
the input COORD/ZCORN information with the final set of active
cells.

If true, function 'processGRDECL' will create a new array

    cpnodes

in the 'G.cells' substructure.  The size of 'cpnodes' is

    G.cells.num-by-8

in which the i-th row identifies the original CP nodes of the
i-th active cell.  The vertices within each row are stored in
the same natural ordering as the input ZCORN array (I before J,
and J before K).