Compute Geometry: Enable Corner-Point Calculation if Supported
This commit enables calculating centroids and cell extents based on
the initial corner-point coordinates if supported by the input grid.
At present, the information must be explicitly requested by setting
a flag (option 'CpGeometry') at the call site and even then the CP
based quantities will only be available if the field 'cells.cpnodes'
exists in the input grid.

The actual calculation of these quantities is deferred to the helper
function 'computeCpGeometry'.