Add NNC Support to 'setupOperatorsTPFA'
This commit makes function 'setupOperatorsTPFA' aware of explicit
non-neighbouring connections (NNCs) that may be the result of pinch
processing or an explicit assignment using the NNC keyword in an
ECLIPSE input file.

We assume that, if present, the connections are stored as a
structure named 'nnc' in the input grid and that this structure has
the fields 'cells' and 'trans'.  The former is an array of pairs of
active cells similar to G.faces.neighbors and the latter is the
associate transmissibility.  In this case, we append the new cell
pairs to the 'N' array and the transmissibility values to both of
the 'T' and 'T_all' arrays.  Those then go into forming the other
operators as normal.