initEclipseGrid: Add Support for Explicit NNCs
This commit makes function 'initEclipseGrid' aware of explicit
non-neighbouring connections as entered using the NNC keyword.  If
the input deck has this keyword, then the result grid will have a
substructure named 'nnc' that contains the pairs of active cells in
a field 'cells' and the associate transmissibility value in a field
called 'trans'.