Add Utility to Detect Live Editor Interactive Mode
This commit adds a new utility function

    mrstIsLiveEditorDir

that, based on a directory name, attempts to answer whether that
directory name is likely to be a temporary "Live Editor" or Cell
Mode directory.  The primary use case is in scripts that depend on
extra inputs--e.g. and ECLIPSE file--which is located relative to
the script file itself rather than managed by MRST's dataset manager
through the 'getDatasetPath' function.

Often such scripts will use a strategy that involves MFILENAME:

    this_dir = fileparts(mfilename('fullpath'))
    input = fullfile(this_dir, 'path', 'to', 'input.txt')

which will fail when run in Cell Mode since MFILENAME will not give
a useful response in that case.  If the script knows a likely or
possible fallback location for 'input.txt' it can then choose to
provide this using code like

    this_dir = fileparts(mfilename('fullpath'))
    if mrstIsLiveEditor(this_dir)
       this_dir = fullfile('path', 'to', 'this', 'directory')
    end

    input = fullfile(this_dir, 'path', 'to', 'input.txt')

This is nevertheless a best-effort facility and most of MRST's
internal scripts should be using 'getDatsetPath' instead.