TPFA Operators: Use 'getFaceTransmissibility' From Core
This commit switches 'setupOperatorsTPFA' to calculating face
transmissibilities using the core function

    getFaceTransmissibility

This is a breaking change since the syntax of that function differs
in important ways from the version that used to be in the ad-core
module.  Specifically, the core version does not support a 'deck'
option, but expects that all pertinent non-neighbouring connections
and transmissibility multipliers are already stored in the grid and
rock structures.

Update MRST's existing callers accordingly to not pass the 'deck'
along as an option to 'setupOperatorsTPFA', but continue to provide
a 'deck' option in 'setupOperatorsTPFA'.  If a caller provides an
input deck argument, this syntax will now issue a diagnostic message
and ignore the deck argument.