Promote getFaceTransmissibility to MRST Core
This commit moves the 'getFaceTransmissibility' function from the
'ad-core' module into the main MRST Core package.  This requires
uncoupling the function from other modules-notably deckformat-and
therefore brings a new protocol for handling transmissibility
multipliers.

We still recognise the 'deck' option, but issue a warning that the
syntax is deprecated and will be discontinued in a future version of
MRST.