Construct AD Solver Compatible Incompressible Fluids
This commit tentatively makes the fluid structures created by
the 'incomp' module's

    init*Fluid*

funtions compatible with the general structures expected by our
AD-based solvers.  In particular, the fluid structures now
support queries of the form

    krw = fluid.krW(sw)
    kro = fluid.krO(so)

    muw = fluid.muW(pw)
    muo = fluid.muO(po)

    bw = fluid.bO(pw)
    bo = fluid.bO(po)

to attempt to fit into the AD-based solvers.  The reciprocal
formation volume factor functions return an appropriately sized
array (or AD variable) of the constant value one (1) to implement
the incompressible nature of these fluids.

For backwards compatibility we also preserve the original syntax

   [mu, rho] = fluid.properties()
   [kr, dkr] = fluid.relperm(fluid.saturation(state))

but we expect that we will be able to deprecate and remove this
syntax in a future release of MRST.  In this commit we merely no
longer document that syntax.