Add Exit Code Translation
This commit adds a new MEX-based helper function that attempts
to translate exit codes from the SYSTEM function into human
readable strings.  We are currently limited to the tables built
into C++'s 'system_category()', but this is a reasonable start.

If the target system does not have a C++ compiler we return a
diagnostic message saying so.