Extra Directories: Exclude Editor Temp Directory
Do not pretend that

    FULLFILE(TEMPDIR, 'Editor_*')

is an MRST directory.  This directory was introduced in a recent
version of MATLAB (present in R2020a) and used to store temporary
files created by the Editor (e.g., while processing a script in
cell mode).

Nevertheless, we translate the TEMPDIR output to the 'tempdir'
string for added reaability if a user happens to have a temporary
directory in their path.