Add New Keyword Reader for Fixed Number of Records
This commit adds a new helper function,

    readFixedNumRecords

which complements the existing, more flexible, reader function

    readDefaultedKW

in that it fills a gap not fully covered.  There are keywords,
notably PVTW, which have a fixed number of records but with an
implied copy operation.  For instance

    PVTW
      1  2  3  4  5 /
      /
      /

is supposed to generate three copies of the first record--one for
each PVT region (assuming NTPVT == 3).  If we use readDefaultedKW
here, then we'll correctly input the first record, but terminate
reading upon encountering the second-empty-record.

This function is a special case reader for those kinds of keywords.