simulateScheduleAD: Add optional input argument processOutputFn to process output before saving to file. Details in comments
Function handle to an optional function that processes the simulation output (wellSols, states and reports) before they are stored to state using the output handlers. Allows for storing only data of interest, which is useful when dealing with large models. Changes made to the output by this function are only applied to the data that that is stored, and will not affect what is passed on to the next timestep.