Use Monolithic Models for EOR Polymer Simulation Runs
This commit adds a new option,

    useLegacyModels

(default false) to 'selectModelFromDeck' and its primary caller
'initEclipseProblemAD'.  This stop-gap measure is the least
expensive way of restoring our ability of running the existing
polymer examples in 'ad-eor' in time for the MRST 2020b release.

The long-term solution is probably to port these examples to
use the Generic* models instead.

Suggested by [at]kalie.